requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
langchain==0.3.7
langchain-openai==0.2.8
langgraph==0.2.54
openai==1.54.3
python-dotenv==1.0.1
boto3>=1.35.0
langgraph-dynamodb-checkpoint>=0.2.6
sqlalchemy>=2.0.30
asyncpg>=0.29.0
pgvector>=0.3.2
